Fix __proto__ property emit
The `__proto__` property is a special use case. It can be defined as a
computed property in an object expression, or via `defineProperty`. If
it’s assigned via a non-computed property in an object expression or via
a regular property assignment, even computed, this sets the prototype of
the object instead. This makes the generated code potentially vulnerable
to prototype pollution.